Reimplement Kartverket Flyfoto as a regular tile layer with token proxy
